pyproject.toml
[project]
name = "simpletraderv3"
version = "3.0.0"
description = "S1mpleTrader V3 - Contract-First, Event-Driven Trading Platform"
authors = [
    {name = "S1mpleTrader Team"}
]
requires-python = ">=3.11"
dependencies = [
    "pydantic>=2.5.0",
    "pyyaml>=6.0.1",
    "mcp>=1.0.0",
    "gitpython>=3.1.0",
    "PyGithub>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.21.0",
    "mypy>=1.7.0",
    "ruff>=0.1.6",
    "pylint>=3.0.0",
    "black>=23.9.0",
    "pyright>=1.1.386",
]

[build-system]
requires = ["setuptools>=68.0.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["."]
include = ["backend*", "tests*", "mcp_server*"]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests", "plugins"]
norecursedirs = [".git", ".vscode", "__pycache__", "docs", "results", "source_data", "frontends", "locales"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=backend",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "integration: Integration tests requiring real dependencies",
    "slow: Tests that take more than 5 seconds to complete",
    "manual: Manual tests requiring full environment setup and explicit enablement (RUN_MANUAL_TESTS=1)",
    "asyncio: Mark async tests for pytest-asyncio",
]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_any_generics = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
follow_imports = "normal"
ignore_missing_imports = false

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "ANN", # flake8-annotations
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "DTZ", # flake8-datetimez
    "T10", # flake8-debugger
    "ISC", # flake8-implicit-str-concat
    "RET", # flake8-return
    "SIM", # flake8-simplify
    "ARG", # flake8-unused-arguments
]
ignore = [
    "ANN101", # Missing type annotation for self
    "ANN102", # Missing type annotation for cls
    "ANN401", # Dynamically typed expressions (Any) are disallowed - needed for **kwargs
    "ARG002", # Unused method argument - needed for interface compliance
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = [
    "ANN",  # No type annotations required in tests
    "ARG",  # Unused arguments OK in fixtures
    "N806", # Variable names can be uppercase (MockManager pattern)
    "E501", # Line length OK in tests
    "SIM117", # Nested with OK in tests
    "SIM105", # try-except-pass OK in tests
]

[tool.pyright]
# Pydantic v2 integration - prevents FieldInfo type inference issues
reportFunctionMemberAccess = false

[tool.pylint.MAIN]
init-hook = 'import sys; sys.path.append(".")'
extension-pkg-allow-list = ["pydantic"]

[tool.pylint.FORMAT]
max-line-length = 100

[tool.pylint.messages_control]
disable = [
    "duplicate-code",  # R0801: often inevitable in tests/scaffolding
    "redefined-outer-name",  # W0621: pytest fixtures use parameter injection
]
